refactor : Movie 객체 패키징 수정
