--- app/src/main/java/woowacourse/movie/domain/model/Movie.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.domain.model
-
-class Movie(
-    val title: String,
-    val poster: String,
-    val releaseDate: ScreeningDate,
-    val runningTime: Int,
-)
